rework operations
