+++ src/meson.build
@@ -12,15 +12,17 @@ graphs_sources = [
   'actions.py',
   'add_data_advanced.py',
   'add_equation.py',
+  'calculation.py',
+  'clipboard.py',
   'canvas.py',
   'colorpicker.py',
   'data.py',
   'file_io.py',
   'graphs.py',
-  'item_operations.py',
   'main.py',
   'misc.py',
-  'operation.py',
+  'operations.py',
+  'operation_tools.py',
   'plot_settings.py',
   'plotting_tools.py',
   'preferences.py',
