+++ tox.ini
@@ -8,4 +8,4 @@ inline-quotes = double
 
 per-file-ignores =
 	# As long as numpy should be available in its current state
-	operation.py: F403, F401
+	calculation.py: F403, F401
